Add rh-bugid Jenkins job
This script is used as gate-keeper so that only bugs against the
GlusterFS project can be used for submitting changes. Many developers
working for Red Hat work on Red Hat Storage (based on GlusterFS) and on
the community GlusterFS code. It is not helpful for the community to
have changes sent against RHS bugs that they can not access.

Original-author: Amar Tumballi <amarts@gmail.com>