Add the compare-bug-version-and-git-branch Jenkins job
For being able to track changes, and confirm that they have been fixed
in certain releases, we need a bug per GlusterFS version if the change
will be submitted for different branches.

Only this way, we can change the status of a bug to ON_QA with a alpha
or beta version. It allows us to list all the bugs that have been fixed
for a particular release.

With this new Jenkins job, the bug/version and git/branch will be
checked, and an error is returned when there is a mismatch.

URL:
http://supercolony.gluster.org/pipermail/gluster-devel/2014-August/041922.html
Signed-off-by: Niels de Vos <ndevos@redhat.com>