Add check for bug status in compare-bug-version-and-git-branch.sh
Currently compare-bug-version-and-git-branch smoke test does not
check for the bug status. As a result it is very well possible
that a patch/change submitted against a BZ under CLOSED status
could bypass this test and proceed assuming success. This change
will add an additional check for bug status. Please see the
following link for more details on bug status life cycle:

http://gluster.readthedocs.org/en/latest/Contributors-Guide/Bug-report-Life-Cycle/

Signed-off-by: Anoop C S <anoopcs@redhat.com>