Adding scripts to run the distaf tests in centos-ci
These two scripts are used to run the distaf nightly tests
in centos-ci infra.

run_distaf_centosci.sh: This bash executable creates a virtualenv
and installs the dependencies. Then it requests duffy for servers
and clients. Once it gets the servers and clients it puts them
in a config file which is then used by distaf to start the tests

update_distaf_config.py: This is a pyhton script which reads the
yml files generated by duffy API calls and then gets the server
and client information. It the reads the distaf config file. It
updates the distaf config file with the server and clients. And
then it dupms them in a yaml format again in a config file.

Both files are required to start the job in centos-ci. These scripts
are by no means perfect. This is just a starting point for more
enhancement.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>