centos-ci: add nightly build job
Build the HEAD of the master branch in mock agains epel-7-x86_64. Other
branches and CentOS versions/architectures can be selected with
environment variables.

The RPMs are published on http://artifacts.ci.centos.org/gluster/nightly

Other tests can monitor the repomd.xml file on the artifacts webserver
and use that as trigger to start testing. For each branch there is a
.repo file under http://artifacts.ci.centos.org/gluster/nightly/ . This
file can be used to install the nightly builds. On CentOS one would do
this:

  # yum -y install centos-release-gluser yum-utils
  # yum-config-manager --add-repo=http://artifacts.ci.centos.org/gluster/nightly/master.repo
  # yum -y install glusterfs-server

Signed-off-by: Niels de Vos <ndevos@redhat.com>