centos-ci/heketi-functional: set ghprbPullId to empty string if unset
In case ghprbPullId is not set in the environment, Python returns None
with os.getenv(). This then translates to the string "None" and the
run-tests.sh script tries to use that as a pull-request number.

Assigning an empty sting in case the envionment variable is not set,
prevents any further issues.

Signed-off-by: Niels de Vos <ndevos@redhat.com>