Add issue filtering by looking at the older commit message
Currently if a new commit message is posted, the github issue
is updated with the same review URL. This is noisy.

This change, looks at the older commit message (if this is not
the first commit) and filters out issues referenced in the
older commit message, with the ones in the newer commit message
thereby reducing duplicate posts to the same github issue.

This is tested using the dry-run method only (obviously).

Signed-off-by: ShyamsundarR <srangana@redhat.com>