centos-ci/gluster-kubernetes: run the job once a day
We currently only run the job on GitHub PullRequests. But we consume
components from other projects, so it is better to also run
periodically. Once a day should be sufficient, and we're nowhere near
the limit of resources we can consume.

By default the scripts already clone/checkout the master branch, with
the PR as option. No further changes to the scripts are needed atm.

Suggested-by: John Mulligan <jmulliga@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>