Add glusterd2 tests
These tests are intended to be run as non-voting job (for now) against
glusterfs Gerrit changes. It has been verified that these tests pass on
a centos 7 machine.

This needs to be a non-voting job as we have one test (TestBitrot) that
fails spuriously. Once that is resolved, we can make this a voting job.

Signed-off-by: Prashanth Pai <ppai@redhat.com>