Clean the workspace as well
Default is wipe-workspace: true in the jenkins jobs. It would wipe out the workspace before any build starts.
In recent cases, it was doing the same but the folders have been created at the time of running test suites by the root user
in the workspace was not deleting up. So Jenkins user was not able to perform the operation.

eg: https://build.gluster.org/job/gh_centos7-regression/23/console