add 'disable-linux-io_uring' flag for el-7 builds
https://github.com/gluster/glusterfs/pull/2060 fails on el-7 because
liburing-devel is not available on them. Hence adding this flag to our
build script so that the PR passes.

Note: 2060 is what introduces the 'disable-linux-io_uring' flag but we
can add it here even before it is merged because ./configure will
complain, but won't error out if an un-recognized (yet) option is given.

Signed-off-by: Ravishankar N <ravishankar@redhat.com>