regression.sh: use a zram device for /d when available
It has been observed that AWS builder instances are very slow sometimes,
apparently because of a limited IOPS. To avoid this issue, the script
regression.sh has been modified to try to mount /d on a zram device,
which will drastically reduce the number of IOPS sent to AWS disks. If
the zram device cannot be created for some reason, it falls back to the
old loop device.

This change should avoid spurious timeout failures caused by too long
executions, and it should also prevent failures in tests that are
time-dependent.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>